requirements.txt
pydantic-settings>=2.8.1
psycopg2-binary>=2.9.10